SCCModular: Moved is_dimension_constant to symbolic.py and added testing
