+++ loki/expression/symbolic.py
@@ -19,7 +19,7 @@ from loki.tools import as_tuple
 
 __all__ = [
     'is_constant', 'symbolic_op', 'simplify', 'accumulate_polynomial_terms',
-    'Simplification', 'SimplifyMapper'
+    'Simplification', 'SimplifyMapper', 'is_dimension_constant'
 ]
 
 
@@ -56,6 +56,22 @@ def is_constant(expr):
     if is_minus_prefix(expr):
         return is_constant(strip_minus_prefix(expr))
     return pmbl.is_constant(expr)
+
+
+def is_dimension_constant(d):
+    """Establish if a given dimension symbol is a compile-time constant"""
+    if isinstance(d, sym.IntLiteral):
+        return True
+
+    if isinstance(d, sym.RangeIndex):
+        if d.lower:
+            return is_dimension_constant(d.lower) and is_dimension_constant(d.upper)
+        return is_dimension_constant(d.upper)
+
+    if isinstance(d, sym.Scalar) and isinstance(d.initial , sym.IntLiteral):
+        return True
+
+    return False
 
 
 def symbolic_op(expr1, op, expr2):
